Change mz in scan peak and precursorMZ to double from float
Change data file format, change what is returned from web app.
Changed scan file parser code to parse precursorMZ as double from parse
as float.

+++ Scan_File_Processor_Importer/src_main/org/yeastrc/spectral_storage/scan_file_processor/input_scan_file/dto/MzML_MzXmlScan.java
@@ -25,7 +25,7 @@ public class MzML_MzXmlScan { // implements MsScanIn {
     private int scanNumber = -1;
     private int precursorScanNum = -1;
 
-    private float precursorMz;
+    private double precursorMz;
     private byte precursorCharge;
     private float retentionTime;
     
@@ -120,12 +120,12 @@ public class MzML_MzXmlScan { // implements MsScanIn {
 	}
 
 
-	public float getPrecursorMz() {
+	public double getPrecursorMz() {
 		return precursorMz;
 	}
 
 
-	public void setPrecursorMz(float precursorMz) {
+	public void setPrecursorMz(double precursorMz) {
 		this.precursorMz = precursorMz;
 	}
