+++ WebService_Web_App/src_spectral_file/org/yeastrc/spectral_storage/spectral_file_common/spectral_file/storage_files_on_disk/common_dto/data_file/SpectralFile_SingleScanPeak_Common.java
@@ -8,13 +8,13 @@ package org.yeastrc.spectral_storage.spectral_file_common.spectral_file.storage_
  */
 public class SpectralFile_SingleScanPeak_Common {
 
-	private float m_over_Z;
+	private double m_over_Z;
 	private float intensity;
 	
-	public float getM_over_Z() {
+	public double getM_over_Z() {
 		return m_over_Z;
 	}
-	public void setM_over_Z(float m_over_Z) {
+	public void setM_over_Z(double m_over_Z) {
 		this.m_over_Z = m_over_Z;
 	}
 	public float getIntensity() {
