+++ WebService_Web_App/src_spectral_file/org/yeastrc/spectral_storage/spectral_file_common/spectral_file/storage_files_on_disk/version_003/data_file/on_disk_dto/SpectralFile_SingleScan_V_003.java
@@ -25,7 +25,7 @@ public class SpectralFile_SingleScan_V_003 {
 	
 	private int parentScanNumber;
 	private byte precursorCharge;
-	private float precursor_M_Over_Z;
+	private double precursor_M_Over_Z;
 	
 	/**
 	 * Number of Scan Peaks
